+++ src/apps/psdexporter/psdwidget.cpp
@@ -4,7 +4,7 @@
 #include "psdwidget.h"
 #include "ui_psdwidget.h"
 
-#include <QtPsdExporter/qpsdtreeitemmodel.h>
+#include "psdtreeitemmodel.h"
 #include "exportdialog.h"
 
 #include <QtCore/QCryptographicHash>
@@ -25,7 +25,8 @@ public:
     void updateAttributes();
     void applyAttributes();
 
-    QPsdTreeItemModel model;
+    QPsdWidgetTreeItemModel widgetModel;
+    PsdTreeItemModel model;
 
     QString errorMessage;
     QSettings settings;
@@ -40,16 +41,17 @@ PsdWidget::Private::Private(::PsdWidget *parent)
     : q(parent)
 {
     setupUi(q);
+    model.setSourceModel(&widgetModel);
     treeView->setModel(&model);
 
-    connect(&model, &QPsdTreeItemModel::dataChanged, q, [this](const QModelIndex &topLeft, const QModelIndex &bottomRight, const QList<int> &roles) {
-        const auto *model = dynamic_cast<const QPsdTreeItemModel *>(topLeft.model());
+    connect(&model, &PsdTreeItemModel::dataChanged, q, [this](const QModelIndex &topLeft, const QModelIndex &bottomRight, const QList<int> &roles) {
+        const auto *model = dynamic_cast<const PsdTreeItemModel *>(topLeft.model());
         if (model) {
             for (int r = topLeft.row(); r <= bottomRight.row(); r++) {
                 for (int c = topLeft.column(); c <= bottomRight.column(); c++) {
                     QModelIndex index = topLeft.sibling(r, c);
                     if (roles.empty() || roles.contains(Qt::CheckStateRole)) {
-                        if (c == QPsdTreeItemModel::Column::Visible) {
+                        if (c == PsdTreeItemModel::Column::Visible) {
                             psdView->setItemVisible(model->layerId(index), model->isVisible(index));
                         } else {
                             q->setWindowModified(true);
@@ -62,7 +64,7 @@ PsdWidget::Private::Private(::PsdWidget *parent)
     });
 
     connect(treeView, &QTreeView::expanded, q, [this](const QModelIndex &index) {
-        const auto *model = dynamic_cast<const QPsdTreeItemModel *>(index.model());
+        const auto *model = dynamic_cast<const PsdTreeItemModel *>(index.model());
         if (model) {
             const auto lyid = model->layerId(index);
             settings.setValue(u"%1-x"_s.arg(lyid), true);
@@ -70,7 +72,7 @@ PsdWidget::Private::Private(::PsdWidget *parent)
     });
 
     connect(treeView, &QTreeView::collapsed, q, [this](const QModelIndex &index) {
-        const auto *model = dynamic_cast<const QPsdTreeItemModel *>(index.model());
+        const auto *model = dynamic_cast<const PsdTreeItemModel *>(index.model());
         if (model) {
             const auto lyid = model->layerId(index);
             settings.setValue(u"%1-x"_s.arg(lyid), false);
@@ -86,7 +88,7 @@ PsdWidget::Private::Private(::PsdWidget *parent)
         treeView->selectionModel()->select(index, QItemSelectionModel::ClearAndSelect | QItemSelectionModel::Rows);
     });
 
-    connect(&model, &QPsdTreeItemModel::fileInfoChanged, q, [this](const QFileInfo &fileInfo) {
+    connect(&model, &PsdTreeItemModel::fileInfoChanged, q, [this](const QFileInfo &fileInfo) {
         windowTitle = fileInfo.fileName();
         q->setWindowTitle(windowTitle);
     });
@@ -476,7 +478,7 @@ void PsdWidget::load(const QString &fileName)
     };
     traverseTreeView(d->treeView->rootIndex());
 
-    d->psdView->setModel(&d->model);
+    d->psdView->setModel(d->model.widgetModel());
 }
 
 void PsdWidget::reload()
