+++ src/psdexporter/qpsdexporterplugin.h
@@ -37,10 +37,10 @@ public:
     virtual ExportType exportType() const = 0;
     virtual QHash<QString, QString> filters() const { return {}; }
 
-    virtual bool exportTo(const QPsdTreeItemModel *model, const QString &to, const QVariantMap &hint) const = 0;
+    virtual bool exportTo(const QPsdExporterTreeItemModel *model, const QString &to, const QVariantMap &hint) const = 0;
 
-    const QPsdTreeItemModel *model() const;
-    void setModel(const QPsdTreeItemModel *model) const;
+    const QPsdExporterTreeItemModel *model() const;
+    void setModel(const QPsdExporterTreeItemModel *model) const;
 
     static QByteArrayList keys() {
         return QPsdAbstractPlugin::keys<QPsdExporterPlugin>(QPsdExporterFactoryInterface_iid, "psdexporter");
