+++ tests/auto/psdexporter/CMakeLists.txt
@@ -2,4 +2,3 @@
 # SPDX-License-Identifier: LGPL-3.0-only OR GPL-2.0-only OR GPL-3.0-only
 
 add_subdirectory(regression)
-add_subdirectory(indexmap)
