--- tests/auto/psdexporter/indexmap/CMakeLists.txt
@@ -1,14 +1,0 @@
-# Copyright (C) 2025 Signal Slot Inc.
-# SPDX-License-Identifier: LGPL-3.0-only OR GPL-2.0-only OR GPL-3.0-only
-
-qt_internal_add_test(tst_qpsdexporter_indexmap
-    SOURCES
-        tst_qpsdexporter_indexmap.cpp
-        psdexportercompareindexmap.h psdexportercompareindexmap.cpp
-    LIBRARIES
-        Qt::PsdCore
-        Qt::PsdGui
-        Qt::PsdExporter
-        Qt::Test
-        Qt::TestPrivate
-)
