--- tests/auto/psdexporter/indexmap/data/layer_link.psd_
@@ -1,19 +1,0 @@
-{
-    "exports": {
-    },
-    "layers": {
-        "21": {
-            "name": "長方形 1",
-            "native": 0,
-            "type": 1,
-            "visible": true
-        },
-        "24": {
-            "name": "長方形 1",
-            "native": 0,
-            "type": 1,
-            "visible": true
-        }
-    },
-    "qtpsdparser.hint": 1
-}
